Add files via upload
This project implements a simple rule engine that parses logical expressions into an Abstract Syntax Tree (AST) and evaluates these rules against a provided data set. The goal of the project is to demonstrate an understanding of parsing, tree structures, and evaluation of logical conditions in a software engineering context.